Split out iptestcontroller to control test process.
